requirements_ALPRO.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
apex==0.1
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
av==8.0.3
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
backcall==0.2.0
black==22.1.0
blis==0.7.5
Bottleneck==1.3.2
cachetools==5.0.0
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
cloudpickle==2.0.0
cupy-cuda112==10.1.0
cycler==0.11.0
cymem==2.0.6
cytoolz==0.11.2
decorator==5.1.1
decord==0.6.0
easydict==1.9
einops==0.4.0
en-core-web-sm==3.2.0
et-xmlfile==1.1.0
executing==0.8.2
fastrlock==0.8
filelock==3.4.2
fonttools==4.29.1
frozenlist==1.3.0
fsspec==2022.1.0
future==0.18.2
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grpcio==1.43.0
horovod==0.23.0
huggingface-hub==0.4.0
idna==3.3
importlib-metadata==4.10.1
ipdb==0.13.9
ipython==8.0.1
isodate==0.6.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
langcodes==3.3.0
lmdb==0.97
lz4==2.1.9
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
msgpack==1.0.3
msgpack-numpy==0.4.7.1
msrest==0.6.21
multidict==6.0.2
murmurhash==1.0.6
mypy-extensions==0.4.3
numexpr==2.8.1
numpy==1.21.2
oauthlib==3.2.0
olefile==0.46
openai==0.14.0
opencv-python==4.5.5.62
openpyxl==3.0.9
packaging==21.3
pandas==1.3.5
pandas-stubs==1.2.0.47
parso==0.8.3
pathspec==0.9.0
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
Pillow-SIMD==9.0.0.post1
pip==22.0.3
platformdirs==2.4.1
preshed==3.0.6
prompt-toolkit==3.0.26
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycparser==2.21
pydantic==1.8.2
pyDeprecate==0.3.1
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytorch-lightning==1.5.9
pytz==2021.3
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.8.0
sentencepiece==0.1.96
setuptools==59.5.0
six==1.16.0
smart-open==5.2.1
spacy==3.2.2
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
stack-data==0.1.4
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.0
thinc==8.0.13
threadpoolctl==3.1.0
timm==0.5.4
tokenizers==0.10.3
toml==0.10.2
tomli==2.0.0
toolz==0.11.2
torch==1.10.2
torchaudio==0.10.2
torchmetrics==0.7.1
torchvision==0.11.3
tqdm==4.62.3
traitlets==5.1.1
transformers==4.11.3
typer==0.4.0
typing-extensions==3.10.0.2
ujson==5.1.0
urllib3==1.26.8
wasabi==0.9.0
wcwidth==0.2.5
Werkzeug==2.0.2
wheel==0.37.1
yarl==1.7.2
zipp==3.7.0
